refactor(meetings): share browser audio device inspection
Keep platform selectors and audio lifecycle decisions in each plugin while sharing DOM device inspection. Scope shared helpers so older published plugin fragments retain their declarations and behavior on newer hosts.
